fixy a folder moves
